pyproject.toml
[project]
name = "ambrogio"
version = "0.1.8"
description = "An opinionated dev agent who tackles tech debt"
authors = [
    {name = "Raoul Scalise",email = "scaliseraoul00@gmail.com"}
]
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.9"
packages = [{include = "ambrogio"}]
dependencies = [
    "libcst (>=1.6.0,<2.0.0)",
    "interrogate (>=1.7.0,<2.0.0)",
    "litellm (>=1.30.3,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "coverage (>=7.4.1,<8.0.0)",
    "pytest (>=8.0.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
ambrogio = "ambrogio.__main__:main"